Separate ui (urwid components) and runner
